requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
gunicorn==20.1.0
numpy==1.21.1
Pillow==8.3.1
protobuf==3.17.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0